[BUGFIX] Changing of time log entries not possible
When changing the start/end time (or date) of a time log entries
the values do not get updated in the database. The start/endtime
field being used purely for data transfer didn't get converted to
the "real" start/end fields which get persisted.

The hotkey handling got added to the Edit.html form.

Resolves: #53