[BUGFIX] Time log entries spanning across midnight
Until now it was not possible to create time log entries which span
across midnight as the end time rolled over to 0:00 of the current
day and even took the start time with it as 0:00 of the current day
is lower than any other time of the day.

The issue is resolved by altering the end time field format to allow
an appended "+1", "+2",  ... value which defines the amount of days
the end field is different from the current day.

Negative day offsets are not allowed for the end time field. The end
time must be larger than the start time in any case. So if a start
time before the current day is wanted just adjust the date field.

Resolves: #56